+++ src/api/routes/v1/user.route.ts
@@ -186,6 +186,7 @@ router.route('/:userId/notes').get(authorize(LOGGED_USER), controller.listUserNo
 router.route('/:userId/notes').post(authorize(LOGGED_USER), controller.createNote);
 router.route('/:userId/notes/:noteId').get(authorize(LOGGED_USER), controller.readUserNote);
 router.route('/:userId/notes/:noteId').post(authorize(LOGGED_USER), controller.updateUserNote);
+router.route('/:userId/notes/:noteId/like').post(authorize(LOGGED_USER), controller.likeUserNote);
 router.route('/:userId/notes/:noteId').delete(authorize(LOGGED_USER), controller.deleteUserNote);
 
 module.exports = router;